frontend/backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
azure-ai-documentintelligence
openai
azure-identity